008 Prepare mobile menu
Menu version for mobile devices is prepared ans is available by hamburger.
secondary design works remained

It's good for UX.


## 011 Minimise assets

The app now is a npm project with gulp, watchify.
Assets are minimised in production version.
Due to putting in production environment practically all files are relocated.
Root catalogue structure is also rewritten.

It's good for connection load and time











Signed-off-by: Kiszuriwalilibori <kiszuriwalilibori@poczta.onet.pl>